Implement upgrades and adjust floor, and fix distance units

+++ Bellefeuille-James-Ice-Game-a3/BackGroundTile.cs
@@ -9,12 +9,12 @@ namespace Game10003
 		public Color colTileColour;
 		public BackGroundTile()
 		{ 
-			float Brightness = Random.Float(4, 6); // this sets the brightness
-			int red = (int)Math.Round(Random.Float(28, 31) * Brightness,0); // these will randomly asign a brown floor color to the tile.
+			float Brightness = Random.Float(4, 5); // this sets the brightness
+			int red = (int)Math.Round(Random.Float(26, 28) * Brightness,0); // these will randomly asign a brown floor color to the tile.
 
-			int green = (int)Math.Round(Random.Float(13, 15) * Brightness, 0);
+			int green = (int)Math.Round(Random.Float(11, 12) * Brightness, 0);
 
-			int blue = (int)Math.Round(Random.Float(0, 10) * Brightness / 2,0);
+			int blue = (int)Math.Round(Random.Float(0, 7) * Brightness / 2,0);
 
             colTileColour = new Color(red, green, blue); // this creates a random brown color
 			position = Vector2.Zero;
